requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-socketio>=5.3.0
flask-login>=0.6.3
flask-wtf>=1.2.0
apscheduler>=3.10.0
cryptography>=41.0.0
python-dotenv>=1.0.0
eventlet>=0.35.0
Werkzeug>=3.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
